Move NotificationPublisher to impl package
