+++ app/src/main/AndroidManifest.xml
@@ -64,7 +64,7 @@
         </provider>
 
         <activity android:name=".ui.eventAdministration.EventAdministrationActivity"></activity>
-        <receiver android:name=".notifications.NotificationPublisher"/>
+        <receiver android:name=".repository.impl.NotificationPublisher"/>
 
     </application>
